+++ itachallenge-challenge/src/main/java/com/itachallenge/gamification/config/GamificationMongoConfig.java
@@ -1,0 +1,9 @@
+package com.itachallenge.gamification.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.mongodb.config.EnableReactiveMongoAuditing;
+
+@Configuration
+@EnableReactiveMongoAuditing
+public class GamificationMongoConfig {
+}