+++ itachallenge-challenge/src/test/java/com/itachallenge/challenge/config/dbchangelog/DataBaseRollBackTest.java
@@ -20,7 +20,7 @@ import java.time.Duration;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-@Testcontainers
+@Testcontainers(disabledWithoutDocker = true)
 @SpringBootTest
 class DataBaseRollBackTest {