+++ itachallenge-user/src/main/java/com/itachallenge/user/exception/UnmodificableSolutionException.java
@@ -1,7 +1,0 @@
-package com.itachallenge.user.exception;
-
-public class UnmodificableSolutionException extends RuntimeException {
-    public UnmodificableSolutionException(String message) {
-        super(message);
-    }
-}